config

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -3,4 +3,5 @@
 exec { 'pkill':
   command  => 'pkill -f killmenow',
   path     => ['/usr/bin', '/usr/sbin', '/bin'],
+  onlyif   => '/usr/bin/pgrep -f killmenow',
 }
